requirements.txt
appdirs==1.4.0
dj-database-url==0.4.2
Django==1.10.5
django-cors-headers==2.0.1
djangorestframework==3.5.3
dweepy==0.2.0
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
whitenoise==3.3.0
